+++ src/Symfony/Component/Form/Tests/AbstractTableLayoutTest.php
@@ -336,7 +336,7 @@ abstract class AbstractTableLayoutTest extends AbstractLayoutTest
 
     public function testCsrf()
     {
-        $this->csrfProvider->expects($this->any())
+        $this->csrfTokenGenerator->expects($this->any())
             ->method('generateCsrfToken')
             ->will($this->returnValue('foo&bar'));